prototype/requirements.txt
crewai
langchain-openai
openai
python-dotenv
pyyaml
yfinance
pandas
numpy
fastapi
uvicorn
websockets
pypdf2
pdfplumber

